requirements.txt
python-binance
plotly==5.10.0